requirements.txt
absl-py==0.8.1
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
dominate==2.5.1
entrypoints==0.3
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
gast==0.2.2
get==2019.4.13
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
netron==4.1.6
networkx==2.4
notebook==6.0.2
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.1
post==2019.4.13
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
public==2019.4.13
pyasn1==0.4.7
pyasn1-modules==0.2.7
Pygments==2.4.2
pyparsing==2.4.5
pyrsistent==0.15.5
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.6.0
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.2
Send2Trash==1.5.0
six==1.13.0
sklearn==0.0
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
virtualenv==16.7.7
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0